requirements.txt
glob
pyinstaller
Pillow